refactor code (#10907)
